Move validation wrapper class into MiniLcm
